requirements.txt
streamlit
torch
transformers
sentence-transformers
pillow
scikit-learn
pandas
openpyxl
numpy
